Registar incidente a funcionar

+++ lib/classes/incidente.dart
@@ -4,7 +4,7 @@ class Incidente {
   final DateTime data;
   final TimeOfDay hora;
   final String descricao;
-  final int gravidade;
+  final double gravidade;
 
   Incidente({required this.data,required this.hora,required this.descricao,required this.gravidade});
